test: use pytest-xdist for parallel testing
