+++ tests/rpcs/cfx/filter/test_pending_tx_filter.py
@@ -1,8 +1,10 @@
+import pytest
 from conflux_web3 import Web3
 from conflux_web3.types import Drip
 from tests._test_helpers.type_check import TypeValidator
 
 class TestPendingTxFilter:
+    @pytest.mark.xdist_group(name="account")
     def test_pending_tx_filter(self, moduled_w3: Web3):
         pending_tx_filter_id = moduled_w3.cfx.new_pending_transaction_filter()
         constucted_pending_tx = moduled_w3.cfx.send_transaction({
