+++ tests/rpcs/cfx/test_cfx_account_query_rpcs.py
@@ -5,6 +5,7 @@ from conflux_web3.types import Base32Address, Drip
 from tests._test_helpers.type_check import TypeValidator
 
 class TestAccountQuery:
+    @pytest.mark.xdist_group(name="account")
     def test_get_balance(self, w3: Web3, address: Base32Address):
         balance = w3.cfx.get_balance(address, w3.cfx.epoch_number-5)
         # the balance is supposed to be non-zero
@@ -18,12 +19,14 @@ class TestAccountQuery:
     #     with pytest.raises(TypeError):
     #         w3.cfx.get_balance()
             
+    @pytest.mark.xdist_group(name="account")
     def test_get_staking_balance(self, w3: Web3, address: Base32Address):
         staking_balance = w3.cfx.get_staking_balance(address, w3.cfx.epoch_number-5)
         assert staking_balance >= 0
         assert isinstance(staking_balance, Drip)
         # TODO: use staking balance contract
-        
+
+    @pytest.mark.xdist_group(name="account")
     def test_get_code(self, w3: Web3, contract_address: Base32Address):
         # test different cases
         # contract address / user address
@@ -33,6 +36,7 @@ class TestAccountQuery:
         user_code = w3.cfx.get_code(w3.cfx.account.create().address)
         assert user_code == HexBytes("0x")
     
+    @pytest.mark.xdist_group(name="account")
     def test_get_admin(self, w3: Web3, contract_address: Base32Address):
         # test different cases
         # contract address / user address
@@ -43,6 +47,7 @@ class TestAccountQuery:
         user_admin = w3.cfx.get_admin(random_contract_address)
         assert user_admin is None
         
+    @pytest.mark.xdist_group(name="account")
     def test_get_storage_at(self, w3: Web3, contract_address: Base32Address, use_testnet: bool):
         # TODO: a potential bug in RPC, at present we ignore the testing in local node
         if use_testnet:
@@ -51,6 +56,7 @@ class TestAccountQuery:
         else:
             pass
         
+    @pytest.mark.xdist_group(name="account")
     def test_get_storage_root(self, w3: Web3, contract_address: Base32Address):
         root = w3.cfx.get_storage_root(contract_address, w3.cfx.epoch_number_by_tag("latest_state"))
         TypeValidator.validate_typed_dict(root, "StorageRoot")
@@ -58,26 +64,31 @@ class TestAccountQuery:
         # TODO: check RPC work pattern
         # root = w3.cfx.get_storage_root(w3.account.create().address)
         # assert not root
-        
+
+    @pytest.mark.xdist_group(name="account")
     def test_get_collateral_for_storage(self, w3: Web3, address: Base32Address):
         storage = w3.cfx.get_collateral_for_storage(address, w3.cfx.epoch_number_by_tag("latest_state"))
         
         assert isinstance(storage, int)
     
+    @pytest.mark.xdist_group(name="account")
     def test_get_sponsor_info(self, w3: Web3, contract_address: Base32Address):
         sponsor_info = w3.cfx.get_sponsor_info(contract_address, w3.cfx.epoch_number_by_tag("latest_state"))
         # assert sponsor_info
         TypeValidator.validate_typed_dict(sponsor_info, "SponsorInfo")
-            
+
+    @pytest.mark.xdist_group(name="account")
     def test_get_account(self, w3: Web3, address: Base32Address):
         account_info = w3.cfx.get_account(address, w3.cfx.epoch_number_by_tag("latest_state"))
         TypeValidator.validate_typed_dict(account_info, "AccountInfo")
 
+    @pytest.mark.xdist_group(name="account")
     def test_get_deposit_list(self, w3:Web3, address: Base32Address):
         deposit_list = w3.cfx.get_deposit_list(address)
         for deposit_info in deposit_list:
             TypeValidator.validate_typed_dict(deposit_info, "DepositInfo")
     
+    @pytest.mark.xdist_group(name="account")
     def test_get_vote_list(self, w3:Web3, address: Base32Address):
         vote_list = w3.cfx.get_vote_list(address, w3.cfx.epoch_number_by_tag("latest_state"))
         for vote_info in vote_list:
