+++ tests/rpcs/cfx/test_cfx_block_rpcs.py
@@ -33,36 +33,42 @@ class TestBlock:
         block_data = preprocess_block_data(block_data, use_testnet)
         return block_data
     
+    @pytest.mark.xdist_group(name="account")
     def test_get_block_by_hash(self, block_data: BlockData, no_full_block_data: BlockData):
         TypeValidator.validate_typed_dict(block_data, "BlockData")
         TypeValidator.validate_typed_dict(no_full_block_data, "BlockData")
 
+    @pytest.mark.xdist_group(name="account")
     def test_get_block_by_epoch_number(self, w3:Web3, block_data: BlockData, use_testnet: bool):
         assert block_data['epochNumber'] is not None
         data_ = w3.cfx.get_block_by_epoch_number(block_data['epochNumber'], True)
         data_ = preprocess_block_data(data_, use_testnet)
         TypeValidator.validate_typed_dict(data_, "BlockData")
-        
+    
+    @pytest.mark.xdist_group(name="account")
     def test_get_block_by_block_number(self, w3:Web3, block_data: BlockData, use_testnet: bool):
         assert block_data['blockNumber'] is not None
         data_ = w3.cfx.get_block_by_block_number(block_data['blockNumber'], True)
         data_ = preprocess_block_data(data_, use_testnet)
         assert dict(data_) == dict(block_data)
 
+    @pytest.mark.xdist_group(name="account")
     def test_get_best_block_hash(self, w3:Web3):
         best_block_hash = w3.cfx.get_best_block_hash()
         assert isinstance(best_block_hash, HexBytes)
         
+    @pytest.mark.xdist_group(name="account")
     def test_get_blocks_by_epoch(self, w3: Web3):
         blocks = w3.cfx.get_blocks_by_epoch("latest_state")
         for block_hash in blocks:
             assert isinstance(block_hash, bytes)
             
+    @pytest.mark.xdist_group(name="account")
     def test_get_skipped_blocks(self, w3: Web3):
         blocks = w3.cfx.get_skipped_blocks_by_epoch("latest_state")
         for block_hash in blocks:
             assert isinstance(block_hash, bytes)
-            
+
     def test_get_blocks_by_hash_with_pivot_assumptions(self, w3: Web3, use_testnet: bool):
         epoch_number = w3.cfx.epoch_number_by_tag("latest_confirmed")
         blocks = w3.cfx.get_blocks_by_epoch(epoch_number)
@@ -73,7 +79,8 @@ class TestBlock:
         )
         block_data = preprocess_block_data(block_data, use_testnet)
         TypeValidator.validate_typed_dict(block_data, "BlockData")
-
+    
+    @pytest.mark.xdist_group(name="account")
     def test_get_block(self, w3: Web3, block_data: BlockData, use_testnet: bool):
         epoch_number = block_data["epochNumber"]
         assert epoch_number is not None
@@ -84,6 +91,7 @@ class TestBlock:
             block = preprocess_block_data(block, use_testnet)
             TypeValidator.validate_typed_dict(block, "BlockData")
     
+    @pytest.mark.xdist_group(name="account")
     def test_epoch_receipts(self, w3: Web3, block_data: BlockData):
         epoch_number = block_data["epochNumber"]
         assert epoch_number is not None
