+++ tests/rpcs/test_txpool_rpcs.py
@@ -1,5 +1,5 @@
 from conflux_web3 import Web3
 
-def test_next_nonce(w3: Web3, address):
+def test_next_nonce(w3: Web3, address: str):
     nonce = w3.txpool.next_nonce(address)
     assert isinstance(nonce, int)
